Changes to be committed: 	new file:   test/utils/auth.util.ts 	modified:   test/utils/index.ts
Changes not staged for commit:
	modified:   test/features/shared-folder/get-file.e2e-spec.ts
	modified:   test/utils/auth.util.ts

On branch feat/refactor
Your branch is ahead of 'origin/feat/refactor' by 3 commits.
  (use "git push" to publish your local commits)

Changes to be committed:
	new file:   test/utils/auth.util.ts
	modified:   test/utils/index.ts

Changes not staged for commit:
	modified:   test/features/shared-folder/get-file.e2e-spec.ts

On branch feat/refactor
Your branch is ahead of 'origin/feat/refactor' by 3 commits.
  (use "git push" to publish your local commits)

Changes to be committed:
	modified:   test/features/shared-folder/get-file.e2e-spec.ts
	new file:   test/utils/auth.util.ts
	modified:   test/utils/index.ts

Changes not staged for commit:
	modified:   test/utils/auth.util.ts

On branch feat/refactor
Your branch is ahead of 'origin/feat/refactor' by 3 commits.
  (use "git push" to publish your local commits)

Changes to be committed:
	modified:   test/features/shared-folder/get-file.e2e-spec.ts
	new file:   test/utils/auth.util.ts
	modified:   test/utils/index.ts

Changes not staged for commit:
	modified:   test/features/shared-folder/get-file.e2e-spec.ts